Major: Refactoring all code to use the new IContainer model. Major: Separated SQL drivers into their own packages. Major: Created NEsper.Log.NLog due to compatibility issues with common logging on .netstandard 2.0. Major: Modified TypeBuilder to be easier to use statically. Major: Removing generation of nupkg from non-production targets. Major: Moving most of rendering out of CodegenCompilerRoslyn and back to render methods of statement and expressions.  Consider as potential work for v7.0. Major: Changes to regression tests for v6.1.0 Minor: Updating names on all windows (upper to lower) Minor: Updating names on methods / operations (upper to lower) Minor: Removed container support for internal entities. Fix: Several issues in SupportContainer (for all solutions). Fix: Issues with array type widening. Fix: Compilation errors in examples (new pattern)

+++ NEsper/NEsper/core/context/mgr/ContextControllerHashedGetterCRC32Serialized.cs
@@ -64,7 +64,7 @@ namespace com.espertech.esper.core.context.mgr
                 bytes = new byte[0];
             }
 
-            var value = bytes.GetCrc32() % _granularity;
+            var value = (int) (bytes.GetCrc32() % _granularity);
             var result = value;
             if (result >= 0)
             {
